# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateIntegrationRequest:
    properties:
      config:
        type: object
      enabled:
        description: Pointer to detect if provided (defaults to true)
        type: boolean
      name:
        type: string
      scope:
        allOf:
        - $ref: '#/definitions/model.WebhookScope'
        description: Required for generic_webhook
      team_id:
        description: Required when scope=team
        type: string
      type:
        $ref: '#/definitions/model.IntegrationType'
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.IntegrationListResponse:
    properties:
      integrations:
        items:
          $ref: '#/definitions/model.Integration'
        type: array
      total:
        type: integer
    type: object
...
  api.TestIntegrationRequest:
    properties:
      mode:
        description: '"dm" for direct message test'
        type: string
    type: object
  api.TestIntegrationResponse:
    properties:
      message:
        type: string
      ok:
        type: boolean
    type: object
...
  api.UpdateIntegrationRequest:
    properties:
      config:
        type: object
      enabled:
        type: boolean
      name:
        type: string
    type: object
...
  model.Integration:
    properties:
      config:
        type: object
      created_at:
        type: string
      direction:
        $ref: '#/definitions/model.IntegrationDirection'
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      scope:
        $ref: '#/definitions/model.WebhookScope'
      team_id:
        type: string
      type:
        $ref: '#/definitions/model.IntegrationType'
      updated_at:
        type: string
    type: object
  model.IntegrationDirection:
    enum:
    - inbound
    - outbound
    type: string
    x-enum-varnames:
    - IntegrationDirectionInbound
    - IntegrationDirectionOutbound
  model.IntegrationType:
    enum:
    - slack
    - telegram
    - alertmanager_webhook
    - generic_webhook
    type: string
    x-enum-varnames:
    - IntegrationTypeSlack
    - IntegrationTypeTelegram
    - IntegrationTypeAlertmanagerWebhook
    - IntegrationTypeGenericWebhook
...
  model.WebhookScope:
    enum:
    - global
    - team
    type: string
    x-enum-varnames:
    - WebhookScopeGlobal
    - WebhookScopeTeam
info:
  contact: {}
  description: TokayOps API
  title: TokayOps API
  version: "1.0"
paths:
...
  /api/v1/integrations:
    get:
      description: Get integrations visible to the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IntegrationListResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List integrations
      tags:
      - integrations
    post:
      consumes:
      - application/json
      description: Create a new integration (admin only)
      parameters:
      - description: Integration data
        in: body
        name: integration
        required: true
        schema:
          $ref: '#/definitions/api.CreateIntegrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Integration'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create integration
      tags:
      - integrations
  /api/v1/integrations/{id}:
    delete:
      description: Delete an integration (admin only)
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete integration
      tags:
      - integrations
    get:
      description: Get an integration by ID
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Integration'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get integration
      tags:
      - integrations
    put:
      consumes:
      - application/json
      description: Update an integration (admin only). Empty secrets keep existing
        values.
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      - description: Integration data
        in: body
        name: integration
        required: true
        schema:
          $ref: '#/definitions/api.UpdateIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Integration'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update integration
      tags:
      - integrations
  /api/v1/integrations/{id}/test:
    post:
      consumes:
      - application/json
      description: Send a test message via the integration (admin only)
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      - description: Test request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.TestIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TestIntegrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Test integration
      tags:
      - integrations
...
swagger: "2.0"